refactor cs converter
