+++ tools/a2mochi/x/cs/convert.go
@@ -3,32 +3,16 @@
 package cs
 
 import (
-	"embed"
-	"encoding/json"
 	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"runtime"
 	"strings"
-	"sync"
-	"time"
 
 	"mochi/ast"
 	"mochi/parser"
 )
 
-var (
-	helperOnce sync.Once
-	helperPath string
-	helperErr  error
-	//go:embed internal/AstJson.cs internal/AstJson.csproj
-	helperFS embed.FS
-)
-
 func stripLong(s string) string { return strings.ReplaceAll(s, "L", "") }
 
-func convertExpr(s string) string {
+func rewriteExpr(s string) string {
 	s = strings.TrimSpace(s)
 	if strings.Contains(s, "string.Join(") {
 		s = strings.Replace(s, "string.Join", "join", 1)
@@ -50,91 +34,6 @@ func convertExpr(s string) string {
 		}
 	}
 	return s
-}
-
-// AST represents a very small subset of a C# source file.
-// Program represents a parsed C# source file in a simplified form.
-type Program struct {
-	Types []Type
-	Src   string
-}
-
-var version string
-
-func init() {
-	_, file, _, _ := runtime.Caller(0)
-	root := filepath.Join(filepath.Dir(file), "../../../..")
-	if b, err := os.ReadFile(filepath.Join(root, "VERSION")); err == nil {
-		version = strings.TrimSpace(string(b))
-	} else {
-		version = "unknown"
-	}
-}
-
-func timestamp() string {
-	loc := time.FixedZone("GMT+7", 7*3600)
-	return time.Now().In(loc).Format("2006-01-02 15:04:05")
-}
-
-type Type struct {
-	Name      string
-	Kind      string
-	Access    string
-	StartLine int
-	EndLine   int
-	Fields    []Field
-	Methods   []Func
-	Doc       string
-}
-
-type Field struct {
-	Name   string
-	Type   string
-	Access string
-	Line   int
-	Doc    string
-}
-
-type Func struct {
-	Name      string
-	Params    []Param
-	Ret       string
-	Access    string
-	Static    bool
-	Body      []string
-	StartLine int
-	EndLine   int
-	Doc       string
-}
-
-type Param struct {
-	Name string
-	Type string
-}
-
-// Parse parses C# source code using the Roslyn compiler.
-func Parse(src string) (*Program, error) { return parseRoslyn(src) }
-
-// ConvertSource converts the parsed AST into Mochi source code.
-func ConvertSource(p *Program) (string, error) {
-	n, err := Convert(p)
-	if err != nil {
-		return "", err
-	}
-	var out strings.Builder
-	fmt.Fprintf(&out, "# Generated by a2mochi cs v%s on %s GMT+7\n", version, timestamp())
-	if p != nil && p.Src != "" {
-		out.WriteString("/*\n")
-		out.WriteString(p.Src)
-		if !strings.HasSuffix(p.Src, "\n") {
-			out.WriteByte('\n')
-		}
-		out.WriteString("*/\n")
-	}
-	if err := ast.Fprint(&out, n); err != nil {
-		return "", err
-	}
-	return out.String(), nil
 }
 
 // Convert converts the parsed Program into a Mochi AST node.
@@ -326,7 +225,7 @@ func convertBodyLines(body []string) []string {
 				l = stripLong(l)
 			}
 		}
-		l = convertExpr(l)
+		l = rewriteExpr(l)
 		out = append(out, l)
 	}
 	return out
@@ -417,73 +316,4 @@ func splitArgs(s string) []string {
 		parts = append(parts, strings.TrimSpace(s[start:]))
 	}
 	return parts
-}
-
-// parseRoslyn invokes the helper C# parser to produce the AST in JSON form.
-func parseRoslyn(src string) (*Program, error) {
-	if err := ensureHelper(); err != nil {
-		return nil, err
-	}
-	tmp, err := os.CreateTemp("", "a2mochi-*.cs")
-	if err != nil {
-		return nil, err
-	}
-	defer os.Remove(tmp.Name())
-	if _, err := tmp.WriteString(src); err != nil {
-		return nil, err
-	}
-	tmp.Close()
-	cmd := exec.Command(helperPath, tmp.Name())
-	out, err := cmd.Output()
-	if err != nil {
-		if ee, ok := err.(*exec.ExitError); ok {
-			return nil, fmt.Errorf("parser error: %s", string(ee.Stderr))
-		}
-		return nil, err
-	}
-	var prog Program
-	if err := json.Unmarshal(out, &prog); err != nil {
-		return nil, err
-	}
-	return &prog, nil
-}
-
-// ensureHelper builds the C# helper program if needed.
-func ensureHelper() error {
-	helperOnce.Do(func() {
-		if _, err := exec.LookPath("dotnet"); err != nil {
-			helperErr = fmt.Errorf("dotnet not installed")
-			return
-		}
-		dir, err := os.MkdirTemp("", "a2mochi-cshelper-")
-		if err != nil {
-			helperErr = err
-			return
-		}
-		csPath := filepath.Join(dir, "AstJson.cs")
-		projPath := filepath.Join(dir, "AstJson.csproj")
-		if data, err := helperFS.ReadFile("internal/AstJson.cs"); err == nil {
-			os.WriteFile(csPath, data, 0644)
-		} else {
-			helperErr = err
-			return
-		}
-		if data, err := helperFS.ReadFile("internal/AstJson.csproj"); err == nil {
-			os.WriteFile(projPath, data, 0644)
-		} else {
-			helperErr = err
-			return
-		}
-		cmd := exec.Command("dotnet", "publish", "-c", "Release", "-o", dir)
-		cmd.Dir = dir
-		if out, err := cmd.CombinedOutput(); err != nil {
-			helperErr = fmt.Errorf("dotnet publish: %v: %s", err, out)
-			return
-		}
-		helperPath = filepath.Join(dir, "AstJson")
-		if runtime.GOOS == "windows" {
-			helperPath += ".exe"
-		}
-	})
-	return helperErr
 }
