+++ tools/a2mochi/x/cs/print.go
@@ -1,0 +1,53 @@
+//go:build slow
+
+package cs
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"runtime"
+	"strings"
+	"time"
+
+	"mochi/ast"
+)
+
+var version string
+
+func init() {
+	_, file, _, _ := runtime.Caller(0)
+	root := filepath.Join(filepath.Dir(file), "../../../..")
+	if b, err := os.ReadFile(filepath.Join(root, "VERSION")); err == nil {
+		version = strings.TrimSpace(string(b))
+	} else {
+		version = "unknown"
+	}
+}
+
+func timestamp() string {
+	loc := time.FixedZone("GMT+7", 7*3600)
+	return time.Now().In(loc).Format("2006-01-02 15:04:05")
+}
+
+// ConvertSource converts the parsed Program into Mochi source code string.
+func ConvertSource(p *Program) (string, error) {
+	n, err := Convert(p)
+	if err != nil {
+		return "", err
+	}
+	var out strings.Builder
+	fmt.Fprintf(&out, "# Generated by a2mochi cs v%s on %s GMT+7\n", version, timestamp())
+	if p != nil && p.Src != "" {
+		out.WriteString("/*\n")
+		out.WriteString(p.Src)
+		if !strings.HasSuffix(p.Src, "\n") {
+			out.WriteByte('\n')
+		}
+		out.WriteString("*/\n")
+	}
+	if err := ast.Fprint(&out, n); err != nil {
+		return "", err
+	}
+	return out.String(), nil
+}
